Floating Sign up box
I have added in the floating sign up box where users are able to enter
their email address which will be added to a mailing list (as comma
separated values saved to a txt file ). The floating box can be expanded
which will then reveal the form. The whole form is in a separate file so
can always be added to any webpage very easily.